+++ R/radaette.R
@@ -17,6 +17,9 @@
 #'
 #' @template return_data.frame
 #'
+#' @importFrom dplyr arrange case_when filter group_by if_else inner_join left_join mutate n rowwise select ungroup
+#' @importFrom magrittr %>%
+#' @importFrom tibble tribble
 #'
 #' @export
 #'
@@ -32,7 +35,7 @@ radaette <- function(ADSL, # nolint
                      lookup = NULL,
                      seed = NULL,
                      na_percentage = 0,
-                     na_vars = list(CNSR = c(NA, 0.1), AVAL = c(1234, 0.1), AVALU = c(1234, 0.1)),
+                     na_vars = list(CNSR = c(NA, 0.1), AVAL = c(1234, 0.1)),
                      cached = FALSE) {
 
   stopifnot(is_logical_single(cached))
@@ -117,6 +120,56 @@ radaette <- function(ADSL, # nolint
     )
   }
 
+  # validate and initialize related variables for Hy's law
+  paramcd_hy <- c("HYSTTEUL", "HYSTTEBL")
+  param_hy <- c("Time to Hy's Law Elevation in relation to ULN", "Time to Hy's Law Elevation in relation to Baseline")
+  param_init_list <- relvar_init(param_hy, paramcd_hy)
+  adsl_hy <- select(ADSL, "STUDYID", "USUBJID", "TRTSDTM", "SITEID", "ARM", "study_duration_secs")
+
+  # create all combinations of unique values in STUDYID, USUBJID, PARAM, AVISIT
+  adaette_hy <- expand.grid( # nolint
+    STUDYID = unique(ADSL$STUDYID),
+    USUBJID = ADSL$USUBJID,
+    PARAM = as.factor(param_init_list$relvar1),
+    stringsAsFactors = FALSE)
+
+  # Add other variables to adaette_hy
+  adaette_hy <- left_join(adaette_hy, adsl_hy, by = c("STUDYID", "USUBJID")) %>% # nolint
+  mutate(
+    PARAMCD = factor(rel_var(
+      df = as.data.frame(adaette_hy),
+      var_values = param_init_list$relvar2,
+      related_var = "PARAM"
+    ))
+  ) %>%
+  mutate(
+    CNSR = sample(c(0, 1), prob = c(0.1, 0.9), size = n(), replace = TRUE),
+    EVNTDESC = if_else(.data$CNSR == 0, "First Post-Baseline Raised ALT or AST Elevation Result", NA_character_),
+    CNSDTDSC = if_else(.data$CNSR == 0, NA_character_,
+      sample(c("Last Post-Baseline ALT or AST Result", "Treatment Start"),
+        prob = c(0.9, 0.1),
+        size = n(), replace = TRUE
+      )
+    )
+  ) %>%
+  rowwise() %>%
+  mutate(ADTM = case_when(
+    .data$CNSDTDSC == "Treatment Start" ~ .data$TRTSDTM,
+    TRUE ~ as.POSIXct(
+      .data$TRTSDTM + sample(seq(0, .data$study_duration_secs), size = n(), replace = TRUE),
+      origin = "1970-01-01"
+    )
+  )) %>%
+  mutate(
+    STARTDT = as.Date(.data$TRTSDTM),
+    ADT = as.Date(.data$ADTM),
+    ADY = as.numeric(.data$ADT - .data$STARTDT + 1),
+    AVAL = .data$ADY / 7,
+    AVALU = "WEEKS"
+  )
+
+  adaette_hy <- select(adaette_hy, -"TRTSDTM", -"study_duration_secs", -"ADT", -"STARTDT")
+
   random_ae_data <- function(lookup_info, patient_info, patient_data) {
     cnsr <- sample(c(0, 1), 1, prob = c(1 - lookup_info$CNSR_P, lookup_info$CNSR_P))
     ae_rep_tte <- patient_data$AVAL[patient_data$PARAMCD == "AEREPTTE"]
@@ -193,6 +246,8 @@ radaette <- function(ADSL, # nolint
     STUDYID = "Study Identifier",
     USUBJID = "Unique Subject Identifier"
   )
+
+  ADAETTE <- rbind(ADAETTE, adaette_hy) #nolint
 
   ADAETTE <- ADSL %>%  #nolint
     dplyr::inner_join(